[RTMD-159] - remove image on /check-your-report page
- Add remove image behaviour
- Get rid of old delete-image behaviour on the /add-image step
